requirements.txt
APScheduler==3.11.1
greenlet==3.2.4
playwright==1.56.0
pyee==13.0.0
python-dotenv==1.2.1
typing_extensions==4.15.0
tzlocal==5.3.1
